requirements.txt
autopep8==1.5.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.1
pycodestyle==2.6.0
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
toml==0.10.2
